requirements.txt
anyio==4.12.1
apsw==3.51.2.0
apswutils==0.1.2
beautifulsoup4==4.14.3
certifi==2026.1.4
click==8.3.1
fastcore==1.12.11
fastlite==0.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
oauthlib==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-fasthtml==0.12.39
python-frontmatter==1.1.0
python-multipart==0.0.22
PyYAML==6.0.3
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
